requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.30.2
azure-functions==1.20.0
azure-mgmt-core==1.4.0
azure-data-tables==12.5.0
azure-storage-blob==12.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
isodate==0.6.1
msrest==0.7.1
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
fonttools==4.52.4
fpdf2==2.5.5
html5lib==1.1
jinja2==3.1.4
MarkupSafe==2.1.5
pdfkit==1.0.0
pillow==10.3.0
pycparser==2.22
pydyf==0.10.0
pypdf2==3.0.1
pyphen==0.15.0
python-barcode==0.15.1
six==1.16.0
tinycss2==1.3.0
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
pyodbc
pytest
xmltodict
zeep
lxml


